fix(#466, #479): add `--quiet` option to suppress excessive log
that may hinder users to understand what uis happening.

It is now possible to give multiple `-q` / `-v` in the command line,
and the verbose level is an algebraci additions of all of them, starting
from INFO level.

BUT if any -v is given, the `Spec.verbosed` trait-parameter is set to
true.